+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## Version 1.7.0
 + Add support for Authentication class
 + Add support for Authentication - Radius
++ Add support for Remote Role Groups
 + Add support for Authentication - LDAP
 
 ## Version 1.6.0